# Improve performance of the 'Create New Node' dialog